+++ frontend/router.js
@@ -51,6 +51,7 @@ const routes = [
     { path: '/registration/:token', name: 'user.registration', component: page('auth/register.vue'), meta: {parentName: 'registration.index'}},
 
     { path: '/dinkes', name: 'dinkes.index', component: page('dinkes/index.vue'), meta: {parentName: 'home'}},
+    { path: '/dinkes/tambah', name: 'dinkes.add', component: page('dinkes/tambah.vue'), meta: {parentName: 'dinkes.index'}},
     { path: '/pengguna/:id', name: 'dinkes.show', component: page('dinkes/show.vue'), meta: {parentName: 'dinkes.index'}},
     { path: '/pengguna/:id/edit', name: 'dinkes.edit', component: page('dinkes/edit.vue'), meta: {parentName: 'dinkes.index'}},
     { path: '/registration-dinkes/:token', name: 'dinkes.registration', component: page('auth/register-dinkes.vue'), meta: {parentName: 'registration.index'}},
